+++ CHANGELOG.md
@@ -4,3 +4,5 @@
 
 - Library became CommonJS compatible.
 - Source code is reorganized to match [original Python library](https://pypi.python.org/pypi/sgp4/).
+- `degreesLat` and `degreesLong` don't adjust input radians value and throw `RangeError` if it's out of bounds
+(breaking change).